Add files via upload
This is the Homework2 file which also serves as the main script. This contains the usage of initializePlate, updateTemperature, and analyzePlate files. It also shows the code for visualization of the initial temperatures, calling the updateTemperature function iteratively saving it as a video, plotting the temperature distributions, and saving it as an image file.